Remove SF2000-specific zero-volume optimisations
I suspect these are causing the loss of sound issues on the latest Multicore build.  All that now remains of the optimisations are the frequency_luts.h lookup table inclusion and a few arithmetic/shift changes in the sound_timer function - both of these should be good to leave in for all platforms.